refactor: unify itemize and enumerate implementations
